Feat/issues 13 (#14)
* feat: [Issues: 13] Add a test watcher to the core package for better workflow

* feat: [Issues: 13] Add support for signed integers and decimal values

---------

+++ packages/core/src/lexer/formatters.ts
@@ -7,7 +7,7 @@ export const stringFormatter = (value: string): string => {
 };
 
 export const numberFormatter = (value: string): number => {
-  return parseInt(value);
+  return parseFloat(value);
 };
 
 export const booleanFormatter = (value: string): boolean => {
